Globenter_Back/backend_ecommerce/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.7
dj-rest-auth==7.0.1
Django==5.2.6
django-allauth==65.11.1
django-cors-headers==4.7.0
django-filter==24.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Faker==33.1.0
filelock==3.16.1
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==10.4.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.16.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.8.2
